refactor(web): move product selection form to its own directory
Follow the convention used by the other navigable forms and give the
product selection form a dedicated directory, splitting the product
option and license acceptance into their own components.

The page keeps rendering the same form; no behavior or visual changes.
